Added support for IG and LI
In this PR, I added support for logging in via Instagram and Linkedin since these two behave differently compared to facebook where it responds with access_token and state as hash parameters. Linkedin returns both as query parameters (it also returns code instead of access_token) and IG returns the access token as a hash parameter but the state is actually a query parameter.